backend/backend/requirements.txt
cairosvg==2.7.1
numpy==1.26.0
opencv-python==4.8.1.78
openai==1.3.0
Pillow==10.1.0
svgpathtools==1.6.1
pypotrace==0.3
svgwrite==1.4.3
lxml==4.9.3
scipy==1.11.3
Flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
opencv-python-headless 
httpx==0.27.2
requests==2.31.0
matplotlib==3.7.2
scipy==1.11.3
scikit-learn==1.3.2
pandas==2.1.1
